+++ docs/en/changes/changes.md
@@ -5,6 +5,7 @@
 #### OAP Server
 
 * Add more entities for Zipkin to improve performance.
+* ElasticSearch: scroll id should be updated when scrolling as it may change.
 
 #### Documentation